Refactor sending emails into a “channel” plugin.
