+++ campaignion_email_to_target/src/Action.php
@@ -5,6 +5,7 @@ namespace Drupal\campaignion_email_to_target;
 use Drupal\campaignion_action\ActionBase;
 use Drupal\campaignion_action\TypeInterface;
 use Drupal\campaignion_email_to_target\Api\Client;
+use Drupal\campaignion_email_to_target\Channel\Email;
 use Drupal\little_helpers\Webform\Submission;
 
 /**
@@ -174,6 +175,13 @@ class Action extends ActionBase {
     }
 
     return $pairs;
+  }
+
+  /**
+   * Get the configured channel.
+   */
+  public function channel() {
+    return new Email();
   }
 
 }
