+++ campaignion_email_to_target/src/Component.php
@@ -6,6 +6,7 @@ use Drupal\little_helpers\Webform\Webform;
 use Drupal\little_helpers\Webform\Submission;
 use Drupal\campaignion_action\Loader;
 
+use Drupal\campaignion_email_to_target\Channel\Email;
 use Drupal\campaignion_email_to_target\Loader as ModeLoader;
 
 /**
@@ -175,63 +176,18 @@ class Component {
   /**
    * Send emails to all selected targets.
    */
-  public function sendEmails($data, Submission $submission) {
-    $nid = $submission->nid;
-    $node = $submission->webform->node;
-    $root_node = $node->tnid ? node_load($node->tnid) : $node;
+  public function sendEmails($data, Submission $submission, $channel) {
     $send_count = 0;
-
     foreach ($data as $serialized) {
       $m = unserialize($serialized);
       $message = new Message($m['message']);
       $message->replaceTokens(NULL, $submission);
       unset($m);
 
-      // Set the HTML property based on availablity of MIME Mail.
-      $email['html'] = FALSE;
-      // Pass through the theme layer.
-      $t = 'campaignion_email_to_target_mail';
-      $theme_d = ['message' => $message, 'submission' => $submission];
-      $email['message'] = theme([$t, $t . '_' . $nid], $theme_d);
-
-      $email['from'] = $message->from;
-      $email['subject'] = $message->subject;
-
-      $email['headers'] = [
-        'X-Mail-Domain' => variable_get('site_mail_domain', 'supporter.campaignion.org'),
-        'X-Action-UUID' => $root_node->uuid,
-      ];
-
-      // Verify that this submission is not attempting to send any spam hacks.
-      if (_webform_submission_spam_check($message->to, $email['subject'], $email['from'], $email['headers'])) {
-        watchdog('campaignion_email_to_target', 'Possible spam attempt from @remote !message',
-                array('@remote' => ip_address(), '!message' => "<br />\n" . nl2br(htmlentities($email['message']))));
-        drupal_set_message(t('Illegal information. Data not submitted.'), 'error');
-        return FALSE;
-      }
-
-      $language = $GLOBALS['language'];
-      $mail_params = array(
-        'message' => $email['message'],
-        'subject' => $email['subject'],
-        'headers' => $email['headers'],
-        'submission' => $submission,
-        'email' => $email,
-      );
-
-      // Mail the submission.
-      $m = $this->mail($message->to, $language, $mail_params, $email['from']);
-      if ($m['result']) {
+      if ($channel->send($message, $submission)) {
         $send_count += 1;
       }
     }
-  }
-
-  /**
-   * Wrapper for drupal_mail().
-   */
-  protected function mail($to, $language, $mail_params, $from) {
-    return drupal_mail('campaignion_email_to_target', 'email_to_target', $to, $language, $mail_params, $from);
   }
 
 }
